backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
ultralytics==8.3.88
opencv-python-headless==4.11.0.86
Pillow==11.1.0
numpy
pydantic==2.10.6
python-dotenv==1.0.1
uvicorn